+++ subsys/pm/CMakeLists.txt
@@ -1,9 +1,11 @@
 # SPDX-License-Identifier: Apache-2.0
 
 if(CONFIG_PM)
-  zephyr_sources(pm.c policy.c state.c)
+  zephyr_sources(pm.c state.c)
   zephyr_sources_ifdef(CONFIG_PM_STATS pm_stats.c)
 endif()
+
+add_subdirectory(policy)
 
 zephyr_sources_ifdef(CONFIG_PM_DEVICE device.c)
 zephyr_sources_ifdef(CONFIG_PM_DEVICE_RUNTIME device_runtime.c)
