+++ subsys/pm/Kconfig
@@ -19,6 +19,8 @@ config PM
 	  power management subsystem of the number of ticks until the next kernel
 	  timer is due to expire.
 
+rsource "policy/Kconfig"
+
 if PM
 
 module = PM
@@ -50,35 +52,6 @@ config PM_NEED_ALL_DEVICES_IDLE
 	help
 	  When this option is enabled, check that no devices are busy before
 	  entering into system low power mode.
-
-choice PM_POLICY
-	prompt "Idle State Power Management Policy"
-	default PM_POLICY_DEFAULT
-	help
-	  Select the idle state power management policy.
-
-config PM_POLICY_DEFAULT
-	bool "Default PM policy"
-	help
-	  This option selects the default PM policy. Default policy is based
-	  on CPU residency times and other constraints imposed by the drivers or
-	  application.
-
-config PM_POLICY_CUSTOM
-	bool "Custom PM Policy"
-	help
-	  This options allows applications to override the default policy with
-	  a custom implementation.
-
-endchoice
-
-config PM_POLICY_DEVICE_CONSTRAINTS
-	bool "Power state constraints per device"
-	help
-	  This option allows devices to have a list of power states
-	  that when the system transition to them, cause power loss in the device.
-	  This used to set and release power state constraints when
-	  it is needed by the device.
 
 endif # PM
